requirements.txt
nextcord
python-dotenv
gamercon-async
python-a2s
